Struct dit_core::peer::Controller

pub struct Controller {
    config: Arc<Config>,
    query_sender: Sender<Query>,
    event_sender: UnboundedSender<Event>,
    shutdown_receiver: Receiver<bool>,
}

Fields
config: Arc<Config>query_sender: Sender<Query>event_sender: UnboundedSender<Event>shutdown_receiver: Receiver<bool>
Implementations

impl Controller

pub async fn closed(&self)

Completes when the connection to the LocalPeer has been closed.

pub async fn shutdown(self) -> Response<()>

Signals the LocalPeer to shut down and completes after all Controllers have been dropped.

pub fn is_graceful_shutdown(&self) -> bool

Returns whether Self::shutdown has been called on any Controller belonging to the LocalPeer.

pub async fn bootstrap(&self, bootstrap_addr: SocketAddr) -> Result<()>

Bootstraps the LocalPeer by sending a GetNeighbors packet to its own SocketAddr.

pub async fn announce(&self, dht_addr: DhtAddr) -> Result<()>

Stores the LocalPeer’s SocketAddr in the DHT at DhtAddr.

async fn query_remote_connect( &self, socket_addr: SocketAddr ) -> Response<ConnectionId>

async fn query_get_neighbors(&self) -> Response<Neighbors>

async fn query_update_neighbors(&self, addrs: DhtAndSocketAddr) -> Response<()>

async fn query_notify_subscribers(&self, packet: Packet) -> Response<()>

async fn query_subscribe( &self, addr: DhtAddr, kind: PayloadKind ) -> Response<Subscription>

fn send_event(&self, event: Event) -> Response<()>

async fn send_query<T>( &self, mk_query: impl FnOnce(Sender<T>) -> Query ) -> Response<T>

async fn connect( &self, socket_addr: SocketAddr ) -> Result<Framed<TcpStream, Codec>>

async fn send_packet( &self, framed: &mut Framed<TcpStream, Codec>, src: DhtAddr, dst: DhtAddr, payload: Payload ) -> Result<()>